Fix remaining Arrow API test failures

+++ fastmob/core/tours_dataframe.py
@@ -136,7 +136,7 @@ class Tours(BaseDataFrame):
             )
         else:
             uid_code_col = "__fastmob_uid_codes__"
-            tl_nw = tl_nw.with_columns(nw.lit(0, dtype=nw.UInt64).alias(uid_code_col))
+            tl_nw = tl_nw.with_columns(nw.lit(0, dtype=nw.UInt32).alias(uid_code_col))
             code_to_uid = {}
 
         (
